+++ include/dynd/dtypes/base_string_dtype.hpp
@@ -44,7 +44,7 @@ public:
 
     // String dtypes stop the iterdata chain
     // TODO: Maybe it should be more flexible?
-    size_t get_iterdata_size(int ndim) const;
+    size_t get_iterdata_size(size_t ndim) const;
 
     void get_dynamic_dtype_properties(const std::pair<std::string, gfunc::callable> **out_properties, size_t *out_count) const;
 };
